requirements.txt
langchain
langchain-community
pypdf
faiss-cpu
openai
tiktoken
fastapi
uvicorn
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
reportlab
chromadb
passlib[bcrypt]
pyjwt
